testing/web-platform/tests/html/browsers/origin/relaxing-the-same-origin-restriction/document_domain_setter_null.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/document_domain_frame.sub.js"></script>
<body>
<!-- SEKRITS! -->
<input id="sekrit" value="omg!">

<script>
  promise_test(t => {
    return createFrame(t)
      .then(f => postMessageToFrame(f, 'poke-at-parent'))
      .then(result => {
        assert_equals(result.data, document.querySelector('#sekrit').value);
        result.frame.remove();
      });
  }, "Access allowed with no 'document.domain' modification. (Sanity check)");

  promise_test(t => {
    return createFrame(t)
      .then(f => postMessageToFrame(f, { domain: null }))
      .then(result => {
        assert_equals(result.data, 'Done');
        return postMessageToFrame(result.frame, 'poke-at-parent')
          .then(result => {
            assert_equals(result.data, 'SecurityError');
            result.frame.remove();
          });
      });
  }, "No access when frame sets a `null` 'document.domain'.");

  promise_test(t => {
    return createFrame(t)
      .then(f => {
        document.domain = null;
        assert_equals(document.domain, "null");
        return postMessageToFrame(f, 'poke-at-parent');
      })
      .then(result => {
        assert_equals(result.data, 'SecurityError');
        result.frame.remove();
      });
  }, "No access when parent sets a `null` 'document.domain'.");

  promise_test(t => {
    return createFrame(t)
      .then(f => {
        document.domain = null;
        assert_equals(document.domain, "null");
        return postMessageToFrame(f, { domain: null });
      })
      .then(result => {
        assert_equals(result.data, 'Done');
        return postMessageToFrame(result.frame, 'poke-at-parent')
          .then(result => {
            assert_equals(result.data, 'SecurityError');
            result.frame.remove();
          });
      });
  }, "No access when both sides set a `null` 'document.domain'.");
</script>
